Move value parser into tailwindcss root
